% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactrend.R
\name{reactrend-shiny}
\alias{reactrend-shiny}
\alias{reactrendOutput}
\alias{renderReactrend}
\alias{reactrend_html}
\title{Shiny bindings for reactrend}
\usage{
reactrendOutput(outputId, width = "100\%\%", height = "400px")

renderReactrend(expr, env = parent.frame(), quoted = FALSE)

reactrend_html(id, style, class, ...)
}
\arguments{
\item{outputId}{output variable to read from.}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a reactrend}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}

\item{id, class}{Valid CSS id and class.}

\item{style}{CSS.}

\item{...}{Any other arguments to pass to the HTML element.}
}
\description{
Output and render functions for using reactrend within Shiny
applications and interactive Rmd documents.
}
