% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactrend.R
\name{reactrend}
\alias{reactrend}
\title{Trend}
\usage{
reactrend(data, draw = FALSE, draw_duration = 3000,
  draw_easing = "ease-in", stroke = "#000", stroke_width = 1,
  stroke_opacity = 1, smooth = FALSE, line_cap = c("butt", "round",
  "square"), radius = 10, padding = 8, dash = NULL,
  dash_offset = 10, gradient = NULL, svg_width = NULL,
  svg_height = NULL, width = NULL, height = "25\%",
  elementId = NULL)
}
\arguments{
\item{data}{A vector of numerical values.}

\item{draw}{Whether to use a drawing animation.}

\item{draw_duration}{Duration of draw animation in milliseconds.}

\item{draw_easing}{Easing to use for draw animation.}

\item{stroke}{Color of trend line.}

\item{stroke_width}{Width of stroke.}

\item{stroke_opacity}{Opacity of stroke.}

\item{smooth}{Whether to smooth trend line.}

\item{line_cap}{Shape of line edges.}

\item{radius}{Controls curve when \code{smooth} is set top \code{TRUE}.}

\item{padding}{Padding for large \code{stroke_width}.}

\item{dash}{A vector of dash length.}

\item{dash_offset}{Controls where dash starts.}

\item{gradient}{Vector of colors to use as gradient.}

\item{svg_width}{Width of svg.}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{elementId}{A valid CSS element id.}
}
\description{
Create a trend line.
}
\examples{
reactrend(runif(100))

reactrend(
  runif(20), 
  gradient = c('#00c6ff', '#F0F', '#FF0'),
  smooth = TRUE,
  draw = TRUE,
  dash = c(1,7,3)
)

}
